# moby-bank-quarkus/src/main/java/com/mobybank/harness/interfaces/rest/ConnectFoldersRequest.java
package com.mobybank.harness.interfaces.rest;

import java.util.List;
import org.eclipse.microprofile.openapi.annotations.media.Schema;

public record ConnectFoldersRequest(
        @Schema(description = "Ids of the folders to connect", required = true) List<String> folderIds) {
}